requirements.txt
beautifulsoup4
gitpython
Jinja2
openai
python-dotenv
python-slugify
